removed the useLogout hook due to over abundance of repeating code, looking to further modularize NavBar, but links have now been modularized. still need to discern why accessTokens are not coming back to the front end
