+++ Frontend/mindFull/src/components/Navbar.tsx
@@ -1,13 +1,22 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
 import styles from '../styles/Navbar.module.css';
 import LogoutModal from './LogoutModal';
-import useLogout from '../hooks/useLogout';
+// import useLogout from '../hooks/useLogout';
 import Notification from './Notification';
+import useNotification from '../hooks/useNotification';
+import NavbarLinks from './NavBarLinks';
+import LogoutButton from '../UI/logoutButton';
+const logoutURL = 'http://localhost:3000/api/logout/';
 
 interface NavBarProps {
   isAuthenticated: boolean;
   setIsAuthenticated: React.Dispatch<React.SetStateAction<boolean>>;
+}
+
+interface HttpError {
+  status: number;
+  message: string;
 }
 
 const NavBar: React.FC<NavBarProps> = ({
@@ -15,14 +24,35 @@ const NavBar: React.FC<NavBarProps> = ({
   setIsAuthenticated,
 }) => {
   const navigate = useNavigate();
-  const {
-    isLogoutModalOpen,
-    openLogoutModal,
-    closeLogoutModal,
-    showNotification,
-    confirmLogout,
-    handleNotification,
-  } = useLogout();
+  const { showNotification, handleNotification } = useNotification();
+  const [isLogoutModalOpen, setIsLogoutModalOpen] = useState(false);
+
+  const openLogoutModal = () => setIsLogoutModalOpen(true);
+  const closeLogoutModal = () => setIsLogoutModalOpen(false);
+
+  const confirmLogout = async () => {
+    try {
+      const response = await fetch(logoutURL, {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        credentials: 'include',
+      });
+
+      if (!response.ok) {
+        throw new Error('Logout failed');
+      }
+
+      setIsAuthenticated(false);
+      closeLogoutModal();
+      handleNotification(3000);
+      navigate('/');
+    } catch (error: unknown) {
+      const knownError = error as HttpError;
+      console.error('Logout failed:', knownError.message);
+    }
+  };
 
   return (
     <nav className={styles.navbar}>
@@ -31,53 +61,19 @@ const NavBar: React.FC<NavBarProps> = ({
         className={styles.navbarBrand}>
         mindFull
       </Link>
-      <ul className={styles.navbarNav}>
-        {!isAuthenticated && (
-          <>
-            <li className={styles.navItem}>
-              <Link to='/login' className={styles.navLink}>
-                Login
-              </Link>
-            </li>
-            <li className={styles.navItem}>
-              <Link to='/signup' className={styles.navLink}>
-                Signup
-              </Link>
-            </li>
-          </>
-        )}
 
-        <li className={styles.navItem}>
-          <Link to='/about' className={styles.navLink}>
-            About
-          </Link>
-        </li>
-        <li className={styles.navItem}>
-          <Link to='/contact' className={styles.navLink}>
-            Contact
-          </Link>
-        </li>
-        <li className={styles.navItem}>
-          <Link to='/resources' className={styles.navLink}>
-            Resources
-          </Link>
-        </li>
-        {showNotification && (
-          <Notification
-            message='You have been successfully logged out.'
-            onClose={() => handleNotification('', 10000)}
-          />
-        )}
-        {isAuthenticated && (
-          <li className={styles.navItem}>
-            <button onClick={openLogoutModal}>Logout</button>
-          </li>
-        )}
-      </ul>
+      <NavbarLinks isAuthenticated={isAuthenticated} />
+
+      {showNotification && (
+        <Notification message='You have been successfully logged out.' />
+      )}
+
+      {isAuthenticated && <LogoutButton onClick={openLogoutModal} />}
+
       {isLogoutModalOpen && (
         <LogoutModal
           closeModal={closeLogoutModal}
-          confirmLogout={() => confirmLogout(navigate, setIsAuthenticated)}
+          confirmLogout={confirmLogout}
         />
       )}
     </nav>
